jpa hibernate 유효성 검증
1. hibernate validate에서 발생하는 오류 수정

- 컬럼 정의 문제

+++ Spring/Globa/src/main/java/org/y2k2/globa/infrastructure/persistence/study/entity/StudyEntity.java
@@ -34,7 +34,7 @@ public class StudyEntity {
     private RecordEntity record;
 
     @Column(name = "study_time", columnDefinition = "INT UNSIGNED DEFAULT 0")
-    private Long studyTime;
+    private Integer studyTime;
 
     @CreationTimestamp
     @Column(name = "created_time", columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
